Fixing review findings: move store_gateway.go and store_gateway_test.go to the storegateway package
